requirements.txt
aiogram~=2.23.1  # Библиотека для разработки Telegram-ботов
aiosqlite~=0.20.0  # Асинхронный интерфейс для SQLite
python-dotenv~=1.0.1  # Загружает переменные окружения из .env файла
pytest~=8.3.2  # Фреймворк для тестирования на Python